Port `/universalis` to Kotlin/JS
