+++ bot/commands/universalis/src/jsMain/kotlin/Handler.kt
@@ -1,0 +1,29 @@
+package cloud.drakon.dynamisbot.universalis
+
+import cloud.drakon.ktdiscord.KtDiscord
+import cloud.drakon.ktuniversalis.KtUniversalis
+import cloud.drakon.ktxivapi.KtXivApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.promise
+import kotlinx.serialization.json.Json
+
+// Initialise these during the initialization phase
+
+val json = Json {
+    ignoreUnknownKeys =
+        true // Not all fields returned by the Discord API are documented
+    isLenient = true // TODO https://github.com/TempestProject/Tempest/issues/3
+}
+
+val ktDiscord = KtDiscord("APPLICATION_ID", "BOT_TOKEN").Interaction("PUBLIC_KEY")
+
+val ktXivApi = KtXivApi
+val ktUniversalis = KtUniversalis
+
+val spanRegex = """<span.*?>|</span>""".toRegex()
+val newLineRegex = """\n{3,}""".toRegex()
+
+@JsExport
+fun handler(event: Any?, context: Any?) = GlobalScope.promise {
+    universalisCommand(json.decodeFromString(event as String))
+}
